Removing TextFieldBase and updating TextField and TextFieldSearch
