+++ ui/components/component-library/component-library-components.scss
@@ -23,7 +23,7 @@
 // Molecules
 @import 'picker-network/picker-network';
 @import 'tag-url/tag-url';
-@import 'text-field-base/text-field-base';
+@import 'text-field/text-field';
 @import 'text-field-search/text-field-search';
 @import 'form-text-field/form-text-field';
 @import 'banner-alert/banner-alert';
