+++ ui/components/component-library/form-text-field/README.mdx
@@ -1,6 +1,6 @@
 import { Story, Canvas, ArgsTable } from '@storybook/addon-docs';
 
-import { TextField, TextFieldBase } from '../';
+import { TextField } from '../';
 import { FormTextField } from './form-text-field';
 
 # FormTextField
@@ -22,10 +22,10 @@ component props
 
 <ArgsTable of={TextField} />
 
-`FormTextField` accepts all [TextFieldBase](/docs/components-componentlibrary-textfield--default-story#props)
+`FormTextField` accepts all [TextField](/docs/components-componentlibrary-textfield--default-story#props)
 component props
 
-<ArgsTable of={TextFieldBase} />
+<ArgsTable of={TextField} />
 
 ### Id
 
@@ -305,7 +305,7 @@ import {
 <FormTextField
   id="custom-spending-cap"
   placeholder="Enter a number"
-  rightAccessory={<ButtonLink>Max</ButtonLink>}
+  endAccessory={<ButtonLink>Max</ButtonLink>}
   marginBottom={4}
   type={TEXT_FIELD_TYPES.NUMBER}
 />
