+++ ui/components/component-library/form-text-field/form-text-field.js
@@ -28,7 +28,7 @@ export const FormTextField = ({
   inputRef,
   label,
   labelProps,
-  leftAccessory,
+  startAccessory,
   maxLength,
   name,
   onBlur,
@@ -37,13 +37,10 @@ export const FormTextField = ({
   placeholder,
   readOnly,
   required,
-  rightAccessory,
+  endAccessory,
   size = Size.MD,
   textFieldProps,
   truncate,
-  showClearButton,
-  clearButtonOnClick,
-  clearButtonProps,
   type = 'text',
   value,
   ...props
@@ -83,7 +80,7 @@ export const FormTextField = ({
         id,
         inputProps,
         inputRef,
-        leftAccessory,
+        startAccessory,
         maxLength,
         name,
         onBlur,
@@ -92,10 +89,7 @@ export const FormTextField = ({
         placeholder,
         readOnly,
         required,
-        rightAccessory,
-        showClearButton,
-        clearButtonOnClick,
-        clearButtonProps,
+        endAccessory,
         size,
         truncate,
         type,
