+++ ui/components/component-library/index.js
@@ -28,11 +28,6 @@ export { Tag } from './tag';
 export { TagUrl } from './tag-url';
 export { Text, TEXT_DIRECTIONS } from './text';
 export { TextField, TEXT_FIELD_TYPES, TEXT_FIELD_SIZES } from './text-field';
-export {
-  TextFieldBase,
-  TEXT_FIELD_BASE_SIZES,
-  TEXT_FIELD_BASE_TYPES,
-} from './text-field-base';
 export { TextFieldSearch } from './text-field-search';
 
 // Molecules
