+++ ui/components/component-library/label/README.mdx
@@ -62,7 +62,7 @@ Use the `htmlFor` prop to allow the `Label` to focus on an input with the same i
 </Canvas>
 
 ```jsx
-import { Label, TextFieldBase } from '../../component-library';
+import { Label, TextField } from '../../component-library';
 
 <Label htmlFor="add-network">Add network</Label>
 <TextField id="add-network" placeholder="Enter network name" />
