--- ui/components/component-library/text-field-base/index.js
@@ -1,5 +1,0 @@
-export { TextFieldBase } from './text-field-base';
-export {
-  TEXT_FIELD_BASE_SIZES,
-  TEXT_FIELD_BASE_TYPES,
-} from './text-field-base.constants';
