--- ui/components/component-library/text-field-base/text-field-base.constants.js
@@ -1,13 +1,0 @@
-import { Size } from '../../../helpers/constants/design-system';
-
-export const TEXT_FIELD_BASE_SIZES = {
-  SM: Size.SM,
-  MD: Size.MD,
-  LG: Size.LG,
-};
-export const TEXT_FIELD_BASE_TYPES = {
-  TEXT: 'text',
-  NUMBER: 'number',
-  PASSWORD: 'password',
-  SEARCH: 'search',
-};
