+++ ui/components/component-library/text-field-search/README.mdx
@@ -1,6 +1,6 @@
 import { Story, Canvas, ArgsTable } from '@storybook/addon-docs';
 
-import { TextFieldBase, TextField } from '..';
+import { TextField } from '..';
 
 import { TextFieldSearch } from './text-field-search';
 
@@ -22,11 +22,6 @@ The `TextFieldSearch` accepts all props below as well as all [Box](/docs/compone
 component props
 
 <ArgsTable of={TextField} />
-
-`TextFieldSearch` accepts all [TextFieldBase](/docs/components-componentlibrary-textfieldbase--default-story#props)
-component props
-
-<ArgsTable of={TextFieldBase} />
 
 ### Clear Button On Click
