+++ ui/components/component-library/text-field-search/text-field-search.js
@@ -6,26 +6,46 @@ import { Size } from '../../../helpers/constants/design-system';
 
 import { ButtonIcon } from '../button-icon';
 import { Icon, ICON_NAMES } from '../icon';
-import { TextFieldBase, TEXT_FIELD_BASE_TYPES } from '../text-field-base';
-import { TextField } from '../text-field';
+import { TextField, TEXT_FIELD_TYPES } from '../text-field';
 
 export const TextFieldSearch = ({
-  value,
-  onChange,
+  className,
+  showClearButton = true, // only works with a controlled input
   clearButtonOnClick,
   clearButtonProps,
-  className,
+  endAccessory,
+  inputProps,
+  value,
+  onChange,
   ...props
 }) => (
   <TextField
     className={classnames('mm-text-field-search', className)}
     value={value}
     onChange={onChange}
-    type={TEXT_FIELD_BASE_TYPES.SEARCH}
-    leftAccessory={<Icon name={ICON_NAMES.SEARCH} size={Size.SM} />}
-    showClearButton
-    clearButtonOnClick={clearButtonOnClick}
-    clearButtonProps={clearButtonProps}
+    type={TEXT_FIELD_TYPES.SEARCH}
+    endAccessory={
+      value && showClearButton ? (
+        <>
+          <ButtonIcon
+            className="mm-text-field__button-clear"
+            ariaLabel="Clear" // TODO: i18n
+            iconName={ICON_NAMES.CLOSE}
+            size={Size.SM}
+            onClick={clearButtonOnClick}
+            {...clearButtonProps}
+          />
+          {endAccessory}
+        </>
+      ) : (
+        endAccessory
+      )
+    }
+    startAccessory={<Icon name={ICON_NAMES.SEARCH} size={Size.SM} />}
+    inputProps={{
+      marginRight: showClearButton ? 6 : 0,
+      ...inputProps,
+    }}
     {...props}
   />
 );
@@ -34,11 +54,16 @@ TextFieldSearch.propTypes = {
   /**
    * The value of the TextFieldSearch
    */
-  value: TextFieldBase.propTypes.value,
+  value: TextField.propTypes.value,
   /**
    * The onChange handler of the TextFieldSearch
    */
-  onChange: TextFieldBase.propTypes.onChange,
+  onChange: TextField.propTypes.onChange,
+  /**
+   * The clear button for the TextFieldSearch.
+   * Defaults to true
+   */
+  showClearButton: PropTypes.bool,
   /**
    * The onClick handler for the clear button
    * Required unless showClearButton is false
@@ -66,6 +91,18 @@ TextFieldSearch.propTypes = {
    * An additional className to apply to the TextFieldSearch
    */
   className: PropTypes.string,
+  /**
+   * Component to appear on the right side of the input
+   */
+  endAccessory: PropTypes.node,
+  /**
+   * Attributes applied to the `input` element.
+   */
+  inputProps: PropTypes.object,
+  /**
+   * FormTextField accepts all the props from TextField and Box
+   */
+  ...TextField.propTypes,
 };
 
 TextFieldSearch.displayName = 'TextFieldSearch';
