+++ ui/components/component-library/text-field/text-field.constants.js
@@ -1,7 +1,13 @@
-import {
-  TEXT_FIELD_BASE_SIZES,
-  TEXT_FIELD_BASE_TYPES,
-} from '../text-field-base/text-field-base.constants';
+import { Size } from '../../../helpers/constants/design-system';
 
-export const TEXT_FIELD_SIZES = TEXT_FIELD_BASE_SIZES;
-export const TEXT_FIELD_TYPES = TEXT_FIELD_BASE_TYPES;
+export const TEXT_FIELD_SIZES = {
+  SM: Size.SM,
+  MD: Size.MD,
+  LG: Size.LG,
+};
+export const TEXT_FIELD_TYPES = {
+  TEXT: 'text',
+  NUMBER: 'number',
+  PASSWORD: 'password',
+  SEARCH: 'search',
+};
